feat: v0.13 — exhaustive audit, 8 critical fixes, 652 tests, production-grade hardening
What changed since v0.12:

Driver (bsql-driver):
- Fixed silent data corruption: decode_date_chrono returned epoch for all dates before 2000-01-01
- Fixed streaming early-drop returning corrupted connections to pool
- Fixed arena offset u32 truncation at >4GB (now usize)
- Fixed i128→i64 truncating cast in OffsetDateTime encode
- Added NULL parameter support (Option<T> Encode with is_null())
- Added Encode for PrimitiveDateTime, NaiveDateTime (TIMESTAMP without tz)
- Added Encode for all array types (bool[], int2[], int4[], int8[], float4[], float8[], text[], bytea[])
- Statement cache auto-invalidation on PG restart (SQLSTATE 26000 re-prepare)
- TCP keepalive via socket2 (60s time, 15s interval)
- Pool: max_lifetime (30min default), acquire_timeout (fail-fast default per CREDO), min_idle, graceful shutdown, PoolStatus metrics
- PoolGuard::drop now detects in-transaction and streaming states
- Negative value validation in array decode and DataRow parse
- constant_time_eq upgraded from u8 to u32 (no length truncation above 255)
- Protocol: explicit CopyIn/CopyOut/CopyBoth error messages, param count validation
- Password zeroization on Config drop
- Unknown sslmode values now error instead of silent default
- Arena single-chunk fast-path in resolve_offset, 64KB stream buffer
- Stack-allocated statement names (no heap alloc on cache miss)

Core (bsql-core):
- Pool: max_lifetime, acquire_timeout, min_idle, close(), PoolStatus with idle/active counts
- Transaction: savepoint/release/rollback_to, IsolationLevel enum (Serializable, RepeatableRead, etc.)
- Stream: advance() method for async row-by-row iteration
- Listener: auto-reconnect with exponential backoff, channel tracking, notification buffer overflow warning, payload size validation, bounded command channel
- Error: from_driver_query (Io mid-query → Query not Connect), DecodeError source chain, is_timeout(), is_serialization_failure()
- Types: JSON, JSONB, INTERVAL, INET, CIDR, MACADDR + array variants

Macros (bsql-macros):
- Fixed TIMESTAMP (OID 1114) decode — was producing compile_error for one of the most common PG types
- Fixed unwrap_or_default() silently returning 0/false/"" on corrupt data — now propagates DecodeError
- Fixed UTF-8 corruption in dynamic query build_variant (byte-by-byte → char-by-char)
- Sort enum SQL fragments validated against injection (semicolons, dangerous keywords)
- E-string literal parsing with backslash escapes
- Semicolon detection for multi-statement rejection
- Dollar-quoted string handling in parameter extraction
- Feature-gated array decode paths (Vec<OffsetDateTime>, Vec<Uuid>, Vec<Decimal>, etc.)
- JSON/JSONB, INTERVAL, INET, CIDR, MACADDR type resolution
- Sort query pre-computed SQL via OnceLock cache (zero runtime allocation)
- Offline auto-fallback: builds without BSQL_DATABASE_URL use .bsql/ cache automatically
- Offline cache format v2 with bsql_version tracking
- Batch column suggestions (single query instead of per-table round-trips)
- Schema-qualified table support in "did you mean?" suggestions
- Rust 2024 keyword `raw` added to reserved list

Testing:
- 652 unit tests (was 390), covering codec edge cases, protocol messages, auth error paths, pool lifecycle, arena boundaries, type mappings, parse edge cases, sort validation

Docs:
- README: removed roadmap, removed (planned) markers for implemented features, version sync
- CREDO: updated doc-test principle with offline auto-fallback clarification
- .bsql/ query cache committed to repo for offline builds out of the box

+++ crates/bsql-driver/tests/integration.rs
@@ -1096,7 +1096,7 @@ async fn streaming_1000_rows() {
 
     loop {
         let num_cols = columns.len();
-        let mut col_offsets: Vec<(u32, i32)> = Vec::new();
+        let mut col_offsets: Vec<(usize, i32)> = Vec::new();
 
         if !first_chunk {
             conn.streaming_send_execute(64).await.unwrap();
@@ -1157,7 +1157,7 @@ async fn streaming_chunk_boundary_exact() {
         .unwrap();
 
     let num_cols = columns.len();
-    let mut col_offsets: Vec<(u32, i32)> = Vec::new();
+    let mut col_offsets: Vec<(usize, i32)> = Vec::new();
     let more = conn
         .streaming_next_chunk(&mut arena, &mut col_offsets)
         .await
@@ -1203,7 +1203,7 @@ async fn streaming_zero_rows() {
         .unwrap();
 
     let num_cols = columns.len();
-    let mut col_offsets: Vec<(u32, i32)> = Vec::new();
+    let mut col_offsets: Vec<(usize, i32)> = Vec::new();
     let more = conn
         .streaming_next_chunk(&mut arena, &mut col_offsets)
         .await
@@ -1234,7 +1234,7 @@ async fn streaming_single_row() {
         .unwrap();
 
     let num_cols = columns.len();
-    let mut col_offsets: Vec<(u32, i32)> = Vec::new();
+    let mut col_offsets: Vec<(usize, i32)> = Vec::new();
     let more = conn
         .streaming_next_chunk(&mut arena, &mut col_offsets)
         .await
@@ -1268,7 +1268,7 @@ async fn streaming_early_drop() {
         .await
         .unwrap();
 
-    let mut col_offsets: Vec<(u32, i32)> = Vec::new();
+    let mut col_offsets: Vec<(usize, i32)> = Vec::new();
     let more = guard
         .streaming_next_chunk(&mut arena, &mut col_offsets)
         .await
